[~TASK] FLOW3 (Package): Added constants for the test directories and made the package manager generate them when creating a package. [~TASK] FLOW3 (Core): The lock manager now avoids an error when no lockfile exists but the site is (supposed to be) locked; a warning is logged instead. Original-Commit-Hash: bafc467cc08b2d71dc1346c71c02e58f64d5679d

+++ TYPO3.Flow/Classes/Core/LockManager.php
@@ -85,7 +85,7 @@ class LockManager {
 		$this->lockPathAndFilename = $this->environment->getPathToTemporaryDirectory() . 'FLOW3.lock';
 		if (file_exists($this->lockPathAndFilename)) {
 			if (filemtime($this->lockPathAndFilename) < (time() - self::LOCKFILE_MAXIMUM_AGE)) {
-				unlink ($this->lockPathAndFilename);;
+				unlink($this->lockPathAndFilename);;
 			} else {
 				$this->siteLocked = TRUE;
 			}
@@ -125,7 +125,11 @@ class LockManager {
 	 */
 	public function unlockSite() {
 		if ($this->siteLocked === TRUE) {
-			unlink ($this->lockPathAndFilename);
+			if (file_exists($this->lockPathAndFilename)) {
+				unlink($this->lockPathAndFilename);
+			} else {
+				$this->systemLogger->log('Site is locked but no lockfile could be found.', LOG_WARNING);
+			}
 			$this->siteLocked = FALSE;
 			$this->systemLogger->log('Unlocked site.', LOG_NOTICE);
 		}
